+++ R/add_id.R
@@ -48,14 +48,14 @@ add_group_id.data.frame <- function(data, ...,
                                     .by = NULL, .cols = NULL,
                                     .name = NULL,
                                     as_qg = FALSE){
-  if (!identical(r_address(.order), r_address(order))){
+  if (!identical(cpp_r_address(.order), cpp_r_address(order))){
     lifecycle::deprecate_warn(
       "0.3.0", what = "add_group_id(order)",
       with = "add_group_id(.order)"
     )
     .order <- order
   }
-  if (!identical(r_address(.ascending), r_address(ascending))){
+  if (!identical(cpp_r_address(.ascending), cpp_r_address(ascending))){
     lifecycle::deprecate_warn(
       "0.3.0", what = "add_group_id(ascending)",
       with = "add_group_id(.ascending)"
@@ -107,8 +107,7 @@ add_group_id.data.frame <- function(data, ...,
     }
   }
   col_to_add <- add_names(list(ids), .name)
-  out <- df_add_cols(df_ungroup(data), col_to_add)
-  reconstruct(data, out)
+  df_add_cols(data, col_to_add)
 }
 #' @rdname add_id
 #' @export
@@ -122,7 +121,7 @@ add_row_id.data.frame <- function(data, ...,
                                   ascending = .ascending,
                                   .by = NULL, .cols = NULL,
                                   .name = NULL){
-  if (!identical(r_address(.ascending), r_address(ascending))){
+  if (!identical(cpp_r_address(.ascending), cpp_r_address(ascending))){
     lifecycle::deprecate_warn(
       "0.3.0", what = "add_row_id(ascending)",
       with = "add_row_id(.ascending)"
@@ -153,7 +152,7 @@ add_row_id.data.frame <- function(data, ...,
         length(group_vars) == length(group_vars(data)) &&
         !groups_changed){
       groups <- group_data(data)
-      o <- unlist(groups[[".rows"]])
+      o <- cpp_unlist_group_locs(groups[[".rows"]])
       sizes <- cheapr::lengths_(groups[[".rows"]])
       row_ids <- cpp_row_id(o, sizes, .ascending)
     } else {
@@ -164,8 +163,7 @@ add_row_id.data.frame <- function(data, ...,
     }
   }
   col_to_add <- add_names(list(row_ids), .name)
-  out <- df_add_cols(df_ungroup(data), col_to_add)
-  reconstruct(data, out)
+  df_add_cols(data, col_to_add)
 }
 #' @rdname add_id
 #' @export
@@ -205,6 +203,5 @@ add_consecutive_id.data.frame <- function(data, ...,
     .name <- unique_col_name(names(data), "consecutive_id")
   }
   col_to_add <- add_names(list(ids), .name)
-  out <- df_add_cols(df_ungroup(data), col_to_add)
-  reconstruct(data, out)
+  df_add_cols(data, col_to_add)
 }