+++ R/f_select.R
@@ -73,7 +73,7 @@ f_select.data.table <- function(data, ..., .cols = NULL){
   keys <- attr(data, "sorted")
   out <- collapse::qDT(out)
   if (all(keys %in% names(out))){
-    if (all(cpp_address_equal(df_select(out, keys), df_select(data, keys)))){
+    if (all(cpp_frame_addresses_equal(df_select(out, keys), df_select(data, keys)))){
       attr(out, "sorted") <- keys
     }
   }
@@ -114,3 +114,6 @@ f_rename.data.table <- function(data, ..., .cols = NULL){
   out <- col_rename(data, .cols = pos)
   collapse::qDT(out)
 }
+#' @rdname f_select
+#' @export
+nothing <- function() character()