+++ R/f_slice.R
@@ -47,12 +47,8 @@
 #' @param replace Should `f_slice_sample()` sample with or without replacement?
 #' Default is `FALSE`, without replacement.
 #' @param weights Probability weights used in `f_slice_sample()`.
-#' @param seed Seed number defining RNG state.
-#' If supplied, this is only applied \bold{locally} within the function
-#' and the seed state isn't retained after sampling.
-#' To clarify, whatever seed state was in place before the function call,
-#' is restored to ensure seed continuity.
-#' If left `NULL` (the default), then the seed is never modified.
+#' @param seed `r lifecycle::badge("deprecated")`
+#' Use `cheapr::with_local_seed()` instead.
 #' @param .order Should the groups be returned in sorted order?
 #' If `FALSE`, this will return the groups in order of first appearance,
 #' and in many cases is faster.
@@ -106,7 +102,7 @@ f_slice <- function(data, i = 0L, ..., .by = NULL,
     if (length(i) == 1 && slice_sign >= 1){
       data_locs <- cheapr::na_rm(list_subset(group_locs, i))
     } else {
-      data_locs <- unlist(cpp_slice_locs(group_locs, i))
+      data_locs <- cpp_unlist_group_locs(cpp_slice_locs(group_locs, i))
     }
     if (is.null(data_locs)){
       data_locs <- integer(0)
@@ -144,7 +140,7 @@ f_slice_head <- function(data, n, prop, .by = NULL,
   } else {
     sequences <- sequence(slice_sizes, from = start, by = 1L)
     if (length(slice_sizes) > 1L){
-      i <- unlist(slice_info[["rows"]], recursive = FALSE, use.names = FALSE)[sequences]
+      i <- cpp_unlist_group_locs(slice_info[["rows"]])[sequences]
     } else {
       i <- sequences
     }
@@ -175,7 +171,7 @@ f_slice_tail <- function(data, n, prop, .by = NULL,
   } else {
     sequences <- sequence(slice_sizes, from = start - slice_sizes + 1L, by = 1L)
     if (length(slice_sizes) > 1L){
-      i <- unlist(slice_info[["rows"]], recursive = FALSE, use.names = FALSE)[sequences]
+      i <- cpp_unlist_group_locs(slice_info[["rows"]])[sequences]
     } else {
       i <- sequences
     }
@@ -291,9 +287,15 @@ f_slice_max <- function(data, order_by, n, prop, .by = NULL,
 #' @rdname f_slice
 #' @export
 f_slice_sample <- function(data, n, replace = FALSE, prop,
-                          .by = NULL, .order = df_group_by_order_default(data),
-                          keep_order = FALSE,
-                          weights = NULL, seed = NULL){
+                           .by = NULL, .order = df_group_by_order_default(data),
+                           keep_order = FALSE,
+                           weights = NULL, seed = NULL){
+  # if (!is.null(seed)){
+  #   lifecycle::deprecate_soft(
+  #     "0.4.0", what = "f_slice_sample(seed)",
+  #     details = "It is recommended to use `cheapr::with_local_seed(f_slice_sample())` instead."
+  #   )
+  # }
   # Check if a seed already exists in global environment
   seed_exists <- exists(".Random.seed")
   # Save it in the first instance
@@ -343,12 +345,12 @@ f_slice_sample <- function(data, n, replace = FALSE, prop,
   } else if (!seed_is_null){
     on.exit({remove(".Random.seed", envir = globalenv())})
   }
-  rows <- unlist(rows, use.names = FALSE, recursive = FALSE)
+  rows <- cpp_unlist_group_locs(rows)
   if (length(rows) > 0L){
     rows <- rows + rep.int(sorted_group_starts(group_sizes, 0L),
                            times = slice_sizes)
   }
-  i <- unlist(slice_info[["rows"]], use.names = FALSE, recursive = FALSE)[rows]
+  i <- cpp_unlist_group_locs(slice_info[["rows"]])[rows]
   if (is.null(i)){
     i <- integer()
   }